requirements.txt
# Dependências principais
google-analytics-data==0.18.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# Web framework
Flask==2.3.3
Flask-CORS==4.0.0

# Visualização de dados
plotly==5.17.0
pandas==2.1.3

# Automação e agendamento
APScheduler==3.10.4

# Email (built-in modules, no need to install)
# smtplib and email are built-in Python modules

# Utilitários
python-dotenv==1.0.0
requests==2.31.0

# IA/OpenRouter
openai==1.3.7

# Google Gemini
google-generativeai==0.3.2

# Excel support
openpyxl==3.1.2

# Slack (opcional - usar webhook ou bot token)
# slack-sdk==3.23.0
